+++ src/script/luna/lunarender.cpp
@@ -354,7 +354,7 @@ void Render::TranslateScreenCoords(double &x, double &y, double w, double h)
         // cross-ref DropBonus in npc_bonus.cpp
         if(g_config.hud_follows_player)
         {
-            int l, t;
+            double l, t;
             if(ScreenType == 5 && !vScreen[2].Visible)
                 GetvScreenAverageCanonical(&l, &t);
             else
